+++ example/android/app/src/main/kotlin/br/storck/tercyo/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package br.storck.tercyo.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()